make node_modules rule non-phony
I believe PHONY recipes are run unconditionally so prerequisites will
not matter and in this case the target node_modules will always run.
Making node_modules non-phony will allow it to only be run when
package.json is updated.

Added clean to as a PHONY target just in case a file named clean might
exist.